Add a distributed setup using GRPC
In this setup the different roles of the processing are running in
different processes. These processes are calling into a task manager
in order to communicate with each other vai state transition (encoded in
the `status` of the `Task` messages).

The clients are mock implementations, and only contain example of the
GRPC calls.

How to try it:

  * Create a venv
  * Install the packages defined in `requirements.txt`
  * Run `build.sh` to generate the message and service stubs
  * Start the task manager first which fires up a GRPC server storing
    tasks
  * Run the client keyboard listener. This contains a mock regarding
    what the keyboard listener should call once the OCR complete
  * Run the llm worker to show an example of how the llm processing
    should happen. Pro tip: To increase parallelism, one can start more
    of these. This takes tasks with `question` and should call the
    LLM to fill out the `answer`.
  * The UI is an example of how to access processed (answered) tasks.